Corrected Pg connection reset on child's death (closes #3063)
The Pg connection of the father was closed by PHP when a child exited (as
the father's Pg connection was inherited by the child from the father's
fork).

The father's connection is now closed before spawning a child, thus
ensuring that the child does not inherit any opened Pg connection from his
father.

A more optimal solution would be to rewrite the process model to ensure a
stricter isolation between the workers and the poller.