fix(ted): Use `ss` when `netstat` is not available (refs #6971)
On a minimal Debian 9, `netstat` is not available by default, but `ss`
(bundled by iproute2) is installed. So, the `ted` rc-init script has
been changed to try `ss` when `netstat` is not available.

If neither `ss` nor `netstat` is available, then the check will assume
the port is not in use.