fix(ted): Add pause at end of `stop` function (refs #6971)
On systemd-enabled systems, a call to `service ted restart` does a
`stop` followed immediately by a `start` (and does not call the
`restart` function), thus bypassing the pause performed by the `restart`
function  between `stop` and `start`.

So, a pause has been introduced at the end of the `stop` function
instead of in the `restart` function.